[fix] 코드 리뷰 후 버그 수정

+++ src/main/java/com/todaybread/server/domain/store/dto/StoreAddRequest.java
@@ -3,6 +3,7 @@ package com.todaybread.server.domain.store.dto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
 import java.sql.Time;
 
 /**
@@ -24,8 +25,8 @@ public record StoreAddRequest (
         @NotBlank String description,
         @NotBlank String addressLine1,
         @NotBlank String addressLine2,
-        @NotNull double latitude,
-        @NotNull double longitude,
+        @NotNull BigDecimal latitude,
+        @NotNull BigDecimal longitude,
         @NotNull Time endTime,
         @NotNull Time lastOrderTime,
         @NotBlank String orderTime
